+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.1.0
+* Uses analytics-android 4.2.6 and Appboy 2.1.2
+  * Appboy 2.1.1 support Android O. Please see our [Android SDK Changelog](https://github.com/Appboy/appboy-android-sdk/blob/master/CHANGELOG.md#added-1) for more information on this.
+* Adds support for custom endpoints which can be set on the Segment dashboard.
+* Fixes a bug where birthdays weren't being sent up to our servers properly. 
+
 ## 2.0.2
 * Fixes a bug where an "Install Attributed" event without any data was causing a crash.